created classes for backend, folders for organization
