+++ backend/database/database_central_system.py
@@ -1,0 +1,44 @@
+import mysql.connector as sql
+
+
+class Databasecentralsystem:
+    def __init__(self, host, user, password, database):
+        self.db = sql.connect(
+            host=host,
+            user=user,
+            password=password,
+            database=hotel
+        )
+        self.cursor = self.db.cursor(dictionary=True)
+
+    def rooms_available(self):
+        self.cursor.execute("SELECT * FROM rooms WHERE Reserved=FALSE")
+        return self.cursor.fetchall()
+
+    def Reservation_save(self, name, email, room_num, checkin, checkout):
+        self.cursor.execute("INSERT INTO Reservation(customer_name, customer_email, room_num, checkin, checkout) VALUES (%s,%s,%s,%s,%s)",
+                            (name, email, room_num, checkin, ckeckout))
+        self.cursor.execute(
+            "UPDATE rooms SET Reserved = TRUE WHERE num=%s", (room_num,))
+        self.db.commit()
+        return self.cursor.lastrowid
+
+    def Cancel_reservation(self, reservation_id):
+        self.cursor.execute(
+            "SELECT room_num FROM Reservation WHERE num=%s", (reservation_id,))
+        result = self.cursor.fetchone()
+        if not result:
+            return False
+
+        room_num = result['room_num']
+
+        self.cursor.execute(
+            "DELETE FROM Reservation WHERE num=%s", (reservation_id,))
+        self.cursor.execute(
+            "UPDATE rooms SET Reserved = FALSE WHERE num=%s", (room_num,))
+        self.db.commit()
+        return True
+
+    def All_reservation(self):
+        self.cursor.execute("SELECT * FROM Reservation")
+        return self.cursor.fetchall()
